v2/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
pymongo
dotenv
pydantic
hashlib
